requirements.txt
google-generativeai>=0.8.4
python-dotenv>=1.0.1
pillow>=11.1.0